Refactored to fix javadoc building
