+++ qupath-core/src/main/java/qupath/lib/projects/ProjectImageEntry.java
@@ -34,6 +34,7 @@ import qupath.lib.images.ImageData;
 import qupath.lib.images.servers.ImageServer;
 import qupath.lib.images.servers.ImageServerBuilder.ServerBuilder;
 import qupath.lib.objects.hierarchy.PathObjectHierarchy;
+import qupath.lib.projects.ProjectResources.ProjectResourceManager;
 
 /**
  * Class to represent an image entry within a project.
